Add comprehensive unit test suite across all 7 crates + GUI frontend

+++ backend/src/handlers/connection.rs
@@ -188,7 +188,7 @@ pub async fn handle_connection(
         netmask: state.network.mask(),
         gateway: state.gateway_ip(),
         dns_server: config.dns,
-        mtu: config.mtu as u16,
+        mtu: config.mtu,
         assigned_ipv6: if ipv6_enabled { Some(assigned_ip6) } else { None },
         netmask_v6: if ipv6_enabled { Some(64) } else { None },
         gateway_v6: if ipv6_enabled { Some(state.gateway_ip_v6()) } else { None },
